False Alarm with installation on Linux
The installation settings for install4j Launcher is set to "Allow only a single running instance of the application." Due to testing crashes, there was a hidden JEquity still running in the background. And running ./JEquity returned nothing while, unbeknownst to me, trying to highlight the background JEquity. So I saw nothing. I added the -verbose switch to the launcher file, redirected to a text file, and, after perusing the output, realized that the Launcher's calling com.install4j.runtime.launcher.util.SingleInstance followed by java.lang.IllegalStateException.
Ah.
So I rebooted the Linux VM, and all was well.